Fixed Chrome bug
When a user is navigating via tab, switches to the mouse, and then switches back to tabbing, the focus ring around <a> elements would not display. I'm not 100% sure WHY, but limiting to controlling the tab focus ONLY at the beginning and end of the cycle seemed to fix the problem.